+++ src/fields/MultiLineText.php
@@ -11,6 +11,7 @@ use verbb\formie\helpers\SchemaHelper;
 use verbb\formie\helpers\StringHelper;
 use verbb\formie\helpers\Variables;
 use verbb\formie\fields\conditions\TextFieldConditionRule;
+use verbb\formie\fields\traits\AutocompleteFieldTrait;
 use verbb\formie\fields\definitions\FieldClientModules;
 use verbb\formie\fields\values\StringFieldValue;
 use verbb\formie\models\ClientModule;
@@ -55,6 +56,12 @@ class MultiLineText extends Field implements SortableFieldInterface, Previewable
     {
         return Schema::TYPE_TEXT;
     }
+
+
+    // Traits
+    // =========================================================================
+
+    use AutocompleteFieldTrait;
 
 
     // Properties
@@ -248,7 +255,7 @@ class MultiLineText extends Field implements SortableFieldInterface, Previewable
 
     public function getSettingGqlTypes(): array
     {
-        return array_merge(parent::getSettingGqlTypes(), [
+        return array_merge(parent::getSettingGqlTypes(), $this->defineAutocompleteGqlType(), [
             'limit' => [
                 'name' => 'limit',
                 'type' => Type::boolean(),
@@ -377,6 +384,9 @@ class MultiLineText extends Field implements SortableFieldInterface, Previewable
                 'includedTypes' => [self::class],
             ]),
             SchemaHelper::prePopulate(),
+            $this->defineAutocompleteSettingSchema([
+                'if' => '!useRichText',
+            ]),
             SchemaHelper::includeInEmailFieldSummariesField(),
             SchemaHelper::lightswitchField([
                 'label' => Craft::t('formie', 'Unique Value'),
@@ -462,6 +472,7 @@ class MultiLineText extends Field implements SortableFieldInterface, Previewable
                     'data-formie-max-chars' => ($this->limit && $this->maxType === 'characters' && $this->max) ? $this->max : null,
                     'data-formie-min-words' => ($this->limit && $this->minType === 'words' && $this->min) ? $this->min : null,
                     'data-formie-max-words' => ($this->limit && $this->maxType === 'words' && $this->max) ? $this->max : null,
+                    'autocomplete' => $this->getAutocompleteCoreAttribute(),
                     'aria-describedby' => $this->instructions ? "{$id}-instructions" : null,
                 ])
                 ->theme([
@@ -507,6 +518,8 @@ class MultiLineText extends Field implements SortableFieldInterface, Previewable
     protected function defineRules(): array
     {
         $rules = parent::defineRules();
+
+        $rules = array_merge($rules, $this->defineAutocompleteRules());
 
         $rules[] = [['min', 'max'], 'number', 'integerOnly' => true];
         $rules[] = [['minType', 'maxType'], 'in', 'range' => ['characters', 'words']];